# Relax type in the initialized array in responsible_map