`responsible_map` was leading to downstream issues where type signature of elements in ensemble solution don't match exactly.

https://github.com/SciML/DiffEqBase.jl/pull/510#issuecomment-629753607
Before

```
julia> @benchmark solve_ode(f, (a = 1, b = 1))
BenchmarkTools.Trial: 
  memory estimate:  6.32 MiB
  allocs estimate:  47533
  --------------
  minimum time:     6.613 ms (0.00% GC)
  median time:      8.372 ms (0.00% GC)
  mean time:        9.322 ms (7.21% GC)
  maximum time:     26.238 ms (0.00% GC)
  --------------
  samples:          535
  evals/sample:     1
 ```
  now
  
 ```
  BenchmarkTools.Trial: 
  memory estimate:  6.32 MiB
  allocs estimate:  47543
  --------------
  minimum time:     6.529 ms (0.00% GC)
  median time:      7.636 ms (0.00% GC)
  mean time:        8.178 ms (7.59% GC)
  maximum time:     20.434 ms (56.60% GC)
  --------------
  samples:          611
  evals/sample:     1
  ```
  
  https://github.com/SciML/DiffEqBase.jl/pull/510#issuecomment-629753801
  
  Before
  ```
  BenchmarkTools.Trial: 
  memory estimate:  796.84 KiB
  allocs estimate:  6016
  --------------
  minimum time:     536.576 μs (0.00% GC)
  median time:      672.692 μs (0.00% GC)
  mean time:        778.878 μs (11.43% GC)
  maximum time:     12.996 ms (88.09% GC)
  --------------
  samples:          6385
  evals/sample:     1
  ```
  now
  
  ```
  BenchmarkTools.Trial: 
  memory estimate:  808.78 KiB
  allocs estimate:  6148
  --------------
  minimum time:     574.533 μs (0.00% GC)
  median time:      726.960 μs (0.00% GC)
  mean time:        862.267 μs (9.92% GC)
  maximum time:     15.025 ms (95.01% GC)
  --------------
  samples:          5749
  evals/sample:     1
  ```