@ChrisRackauckas 2021-04-01
Share the full example?