@Vaibhavdixit02 2021-04-01
```
using OrdinaryDiffEq
function f(du,u,p,t)
  du[1] = p.a
  du[2] = p.b
end

const alg = Tsit5()

function solve_ode(f::F, p::P) where {F,P}

  tspan = (0., 1.0)
  Δt = tspan[2] - tspan[1]
  dt = 1/252
  nodes = Int(ceil(Δt / dt) + 1)
  t = T = [tspan[1] + (i - 1) * dt for i = 1:nodes]

  # if I do not set {true}, prob type Any...
  prob = ODEProblem{true}(f, [0., 0.], tspan, p)
  # prob = ODEProblem(f, [0., 0.], tspan, p)

  prob_func = (prob, i, repeat) -> begin
    remake(prob, tspan = (T[i + 1], t[1]))
  end

  # ensemble problem
  odes = EnsembleProblem(prob, prob_func = prob_func)

  sol = OrdinaryDiffEq.solve(
    odes, OrdinaryDiffEq.Tsit5(), OrdinaryDiffEq.EnsembleThreads(),
    trajectories = nodes - 1, saveat = -dt
  )

  return sol
end
```

```
using DiffEqBase
using DiffEqJump
sr = [1.0,2.0,50.]
maj = MassActionJump(sr,[[1 => 1],[1 => 1],[0 => 1]], [[1 => 1], [1 => -1], [1 => 1]])
params = (1.0,2.0,50.)
tspan = (0.,4.)
u0 = [5]
dprob = DiscreteProblem(u0, tspan, params)
jprob = JumpProblem(dprob, Direct(), maj)
sol = solve(EnsembleProblem(jprob), SSAStepper(), EnsembleThreads(); trajectories=10)
```