@Vaibhavdixit02 2021-04-01
It hits the `responsible_map` I checked that before running the benchmarks