@ChrisRackauckas 2021-04-01
Can you just test EnsembleSerial directly though?