@Vaibhavdixit02 2021-04-01
before
```
julia> @benchmark sol = solve(EnsembleProblem(jprob), SSAStepper(), EnsembleSerial(); trajectories=1000)
BenchmarkTools.Trial: 
  memory estimate:  85.52 MiB
  allocs estimate:  678885
  --------------
  minimum time:     67.045 ms (0.00% GC)
  median time:      97.412 ms (22.76% GC)
  mean time:        97.469 ms (21.78% GC)
  maximum time:     360.689 ms (76.71% GC)
  --------------
  samples:          52
  evals/sample:     1
```

now

```
julia> @benchmark sol = solve(EnsembleProblem(jprob), SSAStepper(), EnsembleSerial(); trajectories=1000)
BenchmarkTools.Trial: 
  memory estimate:  85.57 MiB
  allocs estimate:  679297
  --------------
  minimum time:     67.093 ms (0.00% GC)
  median time:      92.439 ms (22.00% GC)
  mean time:        92.301 ms (20.97% GC)
  maximum time:     313.920 ms (75.61% GC)
  --------------
  samples:          55
  evals/sample:     1
  ```