@ChrisRackauckas 2021-04-01
Can we find a test for this? I assume this will break in the future if we don't capture the Pumas case in a test.